Ditched the or syntax and used more if statements and tried out 'try' syntax
But cutting out the 'or's and using separate 'if's u can tell the user which sum makes the integer they want. Then to keep it all in one loop use the 'try' syntax to allow the code to check the rest of the numbers in the list without crashing because it starts trying to test numbers outside the range of the loop.